requirements.txt
matplotlib==3.2.2
numpy==1.19.2
pyyaml==5.3.1
scikit-learn==0.23.2
tqdm==4.50.2
torch
torchvision
